+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Next
+- feat: Added EasyRefresh.isNested parameter to determine whether to handle NestedScrollView. In versions 3.4.0 and earlier, no setting is required.
 - fix: Adjust the Flutter version, Flutter >=3.10.0.
 
 ## 3.4.0